INS-2644: Move definitions of Methods/Constructors to insolar package (due to cycle dependencies)
